% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMMA.R
\docType{data}
\name{BDMMA}
\alias{BDMMA}
\title{Bayesian Dirichlet--Multinomial approach for meta-analysis of metagenomic read counts}
\usage{
BDMMA(Microbiome_dat, abundance_threshold = 5e-05, burn_in = 5000,
  sample_period = 5000, bFDR = 0.1, PIPcut = 0.5)
}
\arguments{
\item{Microbiome_dat}{A SummarizedExperiment object that includes the taxonomy read counts, phenotypes and batch labels.}

\item{abundance_threshold}{The minimum abundance level for the taxa to be included (default value = 5e-05).}

\item{burn_in}{The length of burn in period before sampling the parameters (default value = 5,000).}

\item{sample_period}{The length of sampling period for estimating parameters' distribution (default value = 5,000)}

\item{bFDR}{The false discovery rate level to control (default value = 0.1).}

\item{PIPcut}{The threshold to cut the posterior inclusion probabilities (PIPs). By default, PIP is thresholding at 0.5.}
}
\value{
A list contains the selected taxa and summary of parameters included in the model.
\item{selected.taxa}{A list includes the selected taxa fesatures that are significantly associated
with the main effect variable.}
\item{parameter_summary}{A data.frame contains the mean and quantiles of the parameters included
in the model. Each row includes a parameter's distribution summary and the parameter name is
labeled in the first row. alpha_g: the baseline intercept of g-th taxon; betaj_g: the association strength between
the g-th taxon and j-th input variables; deltai_g: the batch effect parameter of batch i, taxon g;
L_g: the posterior selection probability of g-th taxon; p: the proportion of significantly associated
taxa; eta: the standard deviation of the spike distribution (in the spike-and-slab prior).}
\item{PIP}{A vector contains the PIPs of selected microbial taxa.}
\item{bFDR}{The corresponding bFDR under the selected microbial taxa.}
}
\description{
Bayesian Dirichlet--Multinomial approach for meta-analysis of metagenomic read counts
}
\examples{
require(SummarizedExperiment)
data(Microbiome_dat)
## (not run)
## output <- BDMMA(Microbiome_dat, burn_in = 3000, sample_period = 3000)
}
\references{
Dai, Zhenwei, et al. "Batch Effects Correction for Microbiome Data with Dirichlet-multinomial Regression." Bioinformatics 1 (2018): 8.
}
