% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L_mean.R
\docType{data}
\name{L_mean}
\alias{L_mean}
\title{Posterior Inclusion Probabilities (PIP)}
\format{A numeric vector including 40 PIP values}
\usage{
L_mean
}
\description{
A dataset containing the posterior inclusion probabilities of 40 variables
}
\keyword{datasets}
