% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDMMA.R
\name{trace_plot}
\alias{trace_plot}
\title{Trace plot of BDMMA output}
\usage{
trace_plot(trace, param, col = "black")
}
\arguments{
\item{trace}{A data.frame named "trace" contained in the output of function BDMMA}

\item{param}{A character vector including the parameters' name for trace_plot}

\item{col}{A string defining the color of trace plot (default color is black)}
}
\value{
The function returns a list containing plot objects of parameters' trace plot.
}
\description{
Trace plot of BDMMA output
}
\examples{
data(dat)
X <- dat$X
Y <- dat$Y
batch <- dat$batch
continuous <- dat$continuous
output <- BDMMA(X, Y, batch, continuous, burn_in = 3000, sample_period = 3000)
figure <- trace_plot(trace, param = c("alpha_1", "beta1_10"))
print(figure)
}
